requirements.txt
streamlit==1.39.0
google-generativeai==0.7.2
python-dotenv